@@ -19,11 +19,11 @@
     "@types/color": "^3.0.2",
     "@types/loadable__component": "^5.13.4",
     "@types/lodash": "^4.14.172",
-    "@types/react": "^17.0.0",
+    "@types/react": "^18.0.17",
     "@types/react-color": "^3.0.6",
-    "@types/react-dom": "^17.0.0",
+    "@types/react-dom": "^18.0.6",
     "@types/react-outside-click-handler": "^1.3.1",
-    "@types/react-router-dom": "^5.1.8",
+    "@types/react-router-dom": "^5.3.3",
     "@types/react-syntax-highlighter": "^13.5.2",
     "@types/styled-components": "^5.1.12",
     "@typescript-eslint/eslint-plugin": "^4.29.1",
@@ -42,11 +42,11 @@
     "lodash": "^4.17.21",
     "markdown-to-jsx": "^7.1.3",
     "prettier": "^2.3.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet-async": "^1.0.9",
     "react-icons": "^4.2.0",
-    "react-router-dom": "^5.2.0",
+    "react-router-dom": "^6.3.0",
     "react-syntax-highlighter": "^15.4.4",
     "styled-components": "^5.3.0",
     "typescript": "^4.3.2",
