@@ -1,8 +1,8 @@
 import { useCallback } from "react";
-import { useHistory, useRouteMatch } from "react-router-dom";
+import { useMatch, useNavigate } from "react-router-dom";
 
 export function useJump(to: string) {
-  const history = useHistory();
-  const match = useRouteMatch({ path: to, exact: true });
-  return useCallback(() => !match && history.push(to), [to, match]);
+  const navigate = useNavigate();
+  const match = useMatch({ path: to, end: true });
+  return useCallback(() => !match && navigate(to), [to, match]);
 }
