@@ -1,6 +1,7 @@
-import { Route, Switch } from "react-router-dom";
+import { Route, Routes } from "react-router-dom";
 import { HelmetProvider } from "react-helmet-async";
 import { AutuiThemeProvider, createTheme } from "autui";
+import { createElement } from "react";
 import Home from "./Home";
 import routes from "./routes";
 import TopNavigation from "./TopNavigation";
@@ -11,12 +12,16 @@ function App() {
   return (
     <HelmetProvider>
       <AutuiThemeProvider initTheme={theme}>
-        <Switch>
+        <Routes>
           {routes.map((it) => (
-            <Route key={it.path} path={it.path} component={it.component} />
+            <Route
+              key={it.path}
+              path={`${it.path}/*`}
+              element={createElement(it.component, {})}
+            />
           ))}
-          <Route path="/" component={Home} />
-        </Switch>
+          <Route path="/" element={<Home />} />
+        </Routes>
         <TopNavigation />
       </AutuiThemeProvider>
     </HelmetProvider>
