@@ -4,17 +4,17 @@ import { IRoute } from "../data/IRoute";
 
 const routes: Array<IRoute> = [
   {
-    path: "/docs",
+    path: "docs",
     title: "Docs",
     component: loadable(() => import("./docs")),
   },
   {
-    path: "/components",
+    path: "components",
     title: "Components",
     component: loadable(() => import("./examples")),
   },
   {
-    path: "/configuration",
+    path: "configuration",
     title: <FaCog />,
     component: loadable(() => import("./Configuartion")),
   },
